feat: Refactor Redux store structure and migrate createRequest and profile slices
