--- app/frontend/src/redux/store.js
@@ -1,10 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import createRequestReducer from './slices/createRequestSlice';
-
-export const store = configureStore({
-  reducer: {
-    createRequest: createRequestReducer,
-  },
-});
-
-export default store;
